Add objects to represent parsed data
I'm adding code/objects to represent columns and rows. Also a small
object to parse US phone numbers and normalize the data.

The `RowSet` object represents a set of rows but it only looks at
columns that correspond to the matching type provided as command
options. It stores the "seen" data in a ruby Set object to do the
heavy lifting of duplicate checking. Wrapped around that core of
a Set is logic to look at all the columns of data in a row, look
at the ones that correspond to a match type, then look those up
in the internal `Set` object.